refactor: Move `symmetrize_widget_helper` and `symmetrize_widget` to `BONEWIDGET_OT_match_symmetrize_shape`, remove `find_match_bones`
Having nested functions like before made changing basically anything unnecessarily complex. In this commit, the functions and the `BONEWIDGET_OT_match_symmetrize_shape` have been refactored in a way that makes the functions obsolete. Ensuring that bones can be mirrored is now handled by the poll method. The description method has been added to help the user understand, why "Symmetrize Shape" is grayed out.

From this commit on, only one bone can be symmetrized at a time. This is, because when trying to symmetrize selected bones, both bones of a pair might be selected, which would lead to unexpected results.