refactor: Move single-use functions from `main_functions.py` to `operators.py`
To improve code readability by avoiding unnecessary code navigation, functions
that are only used once have been moved to the operator classes where they are
used. Syncing to bone transforms is no longer possible, when the widget object
is selected, because the same object can be a widget of multiple bones.